updated for use of repo branch/tag names and branch hash value
